Added framework in Recipe and Ingredient classes
Built a simple Recipe and Ingredient class to get them started. Still
not sure how the SQL is going to be incorporated into this so some of
this may eventually need to be changed or completely removed. Still
getting used to github, so I’m sorry if I messed anything up. Let me
know if I did so that way I can avoid it next time.